+++ packages/common-near/src/project/types.ts
@@ -1,7 +1,7 @@
 // Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
-import {IProjectManifest, ProjectNetworkConfig} from '@subql/common';
+import {IProjectManifest, ProjectNetworkConfig} from '@subql/types-core';
 import {NearDatasource} from '@subql/types-near';
 
 // All of these used to be redefined in this file, re-exporting for simplicity
@@ -26,6 +26,5 @@ export {
 export type INearProjectManifest = IProjectManifest<NearDatasource>;
 
 export interface NearProjectNetworkConfig extends ProjectNetworkConfig {
-  genesisHash?: string;
   chainId?: string;
 }