+++ packages/node/src/indexer/api.service.ts
@@ -1,11 +1,10 @@
 // Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
-import { Inject, Injectable } from '@nestjs/common';
+import { ConnectionInfo } from '@near-js/providers/lib/fetch_json';
+import { Inject, Injectable, OnApplicationShutdown } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import {
-  IndexerEvent,
-  NetworkMetadataPayload,
   getLogger,
   NodeConfig,
   profilerWrap,
@@ -22,7 +21,6 @@ import {
   EpochValidatorInfo,
   GasPrice,
 } from 'near-api-js/lib/providers/provider';
-import { ConnectionInfo } from 'near-api-js/lib/utils/web';
 
 import { SubqueryProject } from '../configure/SubqueryProject';
 import * as NearUtil from '../utils/near';