+++ packages/node/src/indexer/ds-processor.service.ts
@@ -7,7 +7,6 @@ import {
   NearCustomDataSource,
   NearDataSource,
   NearDatasourceProcessor,
-  NearNetworkFilter,
 } from '@subql/common-near';
 import { BaseDsProcessorService } from '@subql/node-core';
 import {
@@ -15,11 +14,13 @@ import {
   SecondLayerHandlerProcessor_1_0_0,
   NearCustomDatasource,
   NearHandlerKind,
+  NearMapping,
+  NearCustomHandler,
 } from '@subql/types-near';
 
 export function isSecondLayerHandlerProcessor_0_0_0<
   K extends NearHandlerKind,
-  F,
+  F extends Record<string, unknown>,
   E,
   DS extends NearCustomDatasource = NearCustomDatasource,
 >(
@@ -33,7 +34,7 @@ export function isSecondLayerHandlerProcessor_0_0_0<
 
 export function isSecondLayerHandlerProcessor_1_0_0<
   K extends NearHandlerKind,
-  F,
+  F extends Record<string, unknown>,
   E,
   DS extends NearCustomDatasource = NearCustomDatasource,
 >(
@@ -46,7 +47,7 @@ export function isSecondLayerHandlerProcessor_1_0_0<
 
 export function asSecondLayerHandlerProcessor_1_0_0<
   K extends NearHandlerKind,
-  F,
+  F extends Record<string, unknown>,
   E,
   DS extends NearCustomDatasource = NearCustomDatasource,
 >(
@@ -77,8 +78,8 @@ export function asSecondLayerHandlerProcessor_1_0_0<
 @Injectable()
 export class DsProcessorService extends BaseDsProcessorService<
   NearDataSource,
-  NearCustomDataSource<string, NearNetworkFilter>,
-  NearDatasourceProcessor<string, NearNetworkFilter>
+  NearCustomDataSource<string, NearMapping<NearCustomHandler>>,
+  NearDatasourceProcessor<string, Record<string, unknown>>
 > {
   protected isCustomDs = isCustomDs;
 }