+++ packages/common-near/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.1] - 2023-09-27
+### Update
+- Update `subql/common` (#64)
+
 ## [2.4.0] - 2023-08-02
 ### Changed
 - update license and `@subql/common` (#56)
@@ -39,7 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add transaction receipts handlers and filters (#13)
 
 ## [1.0.0] - 2023-01-26
-[Unreleased]: https://github.com/subquery/subql-near/compare/common-near/2.4.0...HEAD
+[Unreleased]: https://github.com/subquery/subql-near/compare/common-near/2.4.1...HEAD
+[2.4.1]: https://github.com/subquery/subql-near/compare/common-near/2.4.0...common-near/2.4.1
 [2.4.0]: https://github.com/subquery/subql-near/compare/common-near/2.3.0...common-near/2.4.0
 [2.3.0]: https://github.com/subquery/subql-near/compare/common-near/2.2.1...common-near/2.3.0
 [2.2.1]: https://github.com/subquery/subql-near/compare/common-near/2.2.0...common-near/2.2.1