+++ packages/node/CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.10.2] - 2023-09-27
+### Update
+- Update `subql/common` (#64)
+
+## [2.10.1] - 2023-09-27
+### Added
+- Support for Delegate actions (#62)
+
 ## [2.10.0] - 2023-08-02
 ### Changed
 - Sync with main sdk (#56)
@@ -110,7 +118,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `toJson` method to function args (#9)
 
 ## 1.18.0 - 2023-01-26
-[Unreleased]: https://github.com/subquery/subql-near/compare/node-near/2.10.0...HEAD
+[Unreleased]: https://github.com/subquery/subql-near/compare/node-near/2.10.2...HEAD
+[2.10.2]: https://github.com/subquery/subql-near/compare/node-near/2.10.1...node-near/2.10.2
+[2.10.1]: https://github.com/subquery/subql-near/compare/node-near/2.10.0...node-near/2.10.1
 [2.10.0]: https://github.com/subquery/subql-near/compare/node-near/2.8.0...node-near/2.10.0
 [2.8.0]: https://github.com/subquery/subql-near/compare/node-near/2.6.1...node-near/2.8.0
 [2.6.1]: https://github.com/subquery/subql-near/compare/node-near/2.6.0...node-near/2.6.1